refactor: 디렉터리 구조 변경 및 네이밍 변경
