+++ src/hooks/apis/createStudyRoom.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { API_END_POINT } from '@shared/common';
 import { camel2Under } from '@shared/utils';
-import { ICreateStudyRequest } from '@shared/types';
+import { ICreateStudyRequest } from '@shared/interface';
 
 export default async function createStudyRoom(
   id: number,
